refactor: refactor utils imports
